Renamed folders within SacStateApp. Fixed Virtualized Lists error. Got rid of unnecesscary files in SacStateApp.
